+++ src/main/java/tech/deplant/java4ever/framework/DebugOptions.java
@@ -1,0 +1,4 @@
+package tech.deplant.java4ever.framework;
+
+public record DebugOptions(boolean enabled, long timeout, boolean throwErrors, ContractAbi... treeAbis) {
+}