+++ src/main/java/tech/deplant/java4ever/framework/datatype/Bool.java
@@ -1,7 +1,6 @@
 package tech.deplant.java4ever.framework.datatype;
 
 import com.fasterxml.jackson.annotation.JsonValue;
-import tech.deplant.java4ever.binding.Abi;
 
 public record Bool(Boolean value) implements AbiValue {